libcrypto/tests/tests/Clipperz/Crypto/Usage.html

<!--

Copyright 2008-2011 Clipperz Srl

This file is part of Clipperz's Javascript Crypto Library.
Javascript Crypto Library provides web developers with an extensive
and efficient set of cryptographic functions. The library aims to
obtain maximum execution speed while preserving modularity and
reusability.
For further information about its features and functionalities please
refer to http://www.clipperz.com

* Javascript Crypto Library is free software: you can redistribute
  it and/or modify it under the terms of the GNU Affero General Public
  License as published by the Free Software Foundation, either version
  3 of the License, or (at your option) any later version.

* Javascript Crypto Library is distributed in the hope that it will
  be useful, but WITHOUT ANY WARRANTY; without even the implied
  warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
  See the GNU Affero General Public License for more details.

* You should have received a copy of the GNU Affero General Public
  License along with Javascript Crypto Library.  If not, see
  <http://www.gnu.org/licenses/>.

-->

<html>
<head>
	<title>Clipperz.Crypto.Usage - TEST</title>
	
	<script type="text/javascript" src="../../../../other_libraries/MochiKit/MochiKit.js"></script>
    <script type="text/javascript" src="../../../SimpleTest/SimpleTest.js"></script>
    <link rel="stylesheet" type="text/css" href="../../../SimpleTest/test.css">

    <script type='text/javascript' src='../../../../other_libraries/JSON/json2.js'></script>

    <script type='text/javascript' src='../../../../other_libraries/JSON/json2.js'></script>
    <script type='text/javascript' src='../../../../js/Clipperz/YUI/Utils.js'></script>
	<script type='text/javascript' src='../../../../js/Clipperz/Base.js'></script>
	<script type='text/javascript' src='../../../../js/Clipperz/Logging.js'></script>
	<script type='text/javascript' src='../../../../js/Clipperz/ByteArray.js'></script>

	<script type='text/javascript' src='../../../../js/Clipperz/Crypto/SHA.js'></script>
	<script type='text/javascript' src='../../../../js/Clipperz/Crypto/AES.js'></script>
	<script type='text/javascript' src='../../../../js/Clipperz/Crypto/PRNG.js'></script>

</head>
<body>
<pre id="test">
<script type="text/javascript">
Clipperz.Crypto.PRNG.defaultRandomGenerator().fastEntropyAccumulationForTestingPurpose();

MochiKit.Base.update(Clipperz, {
	'PM': {
		'Crypto': {
			'passwordEntropy': function(aValue) {
				var result;
				var	bitPerChar;

				bitPerChar = 4;
				if (/[a-z]/.test(aValue)) {
					bitPerChar ++;
				}
				if (/[A-Z]/.test(aValue)) {
					bitPerChar ++;
				}
				if (/[^a-zA-Z0-9]/.test(aValue)) {
					bitPerChar ++;
				}
				//MochiKit.Logging.logDebug("--- bitPerChar: " + bitPerChar);

				result = aValue.length * bitPerChar;

				return result;
			}
		}
	}
})

try {
	var keyValue;
	var keyEntropy;
	var	key;
	var plainText;
	var cypherText;
	var randomBytes;
	var hashedValue;
        
        key = Clipperz.Crypto.SHA.sha_d256(new Clipperz.ByteArray("ServerNull1"));
        value = new Clipperz.ByteArray().appendBase64String("nlmchzfX2TEEV3yeoeIoTrpXcZmfQ5/sEaKbY5/N01KiVyYYeHAORXoiMO9tumsj6aH6JKultfJimWP5u3g0c7m5/LxY/bOh6TUnZJkQ0mle8gs0iDKcJfSLy3P1NS+Ulvz0HKsFETepEwVaE4FyBCXMFfUmbyk=");
        plaintext = Clipperz.Crypto.AES.decrypt(key, value);
	alert(plaintext.asString());
        
	//key = Clipperz.Crypto.SHA.sha256(new Clipperz.ByteArray("ServerNull1"));
        //key2 = Clipperz.Crypto.SHA.sha256(new Clipperz.ByteArray("ServerNull2"));
	//keyEntropy = Clipperz.PM.Crypto.passwordEntropy(key);

	//cypherText = Clipperz.Crypto.AES.encrypt(key, new Clipperz.ByteArray("some text to encrypt"));
        
        //cypherText = new Clipperz.ByteArray("i17stxA6IH1I+D3iG1VTywMA4G8yVqaYE1JGiU23jSgATDht2Z7xkiIgpzZkI5DhofXq6QJHldhKXNkgIxaILsHm1XsoX3UvOryqHCRArQefL9cCi/xwxH+0/kpAi95QvBKmbktNirCEvpqmJMHVnQMLTMnHB6+O+RjoTZhcPk4d6ZKMRHGHY6ByXkNTM1OQ5HyJAjkstmNoDOV4S7PTEAJkb4xdHNNVLJyRMxmEPSfCNyLIsPeosiPR11G+Ol2hUPwjuDKZHl6ROGk+5Qfl6HeYdEYM/+5LXITMWRWOhzF1hzGnOo1LwSp7WcqFu39Dz/kwMtvQ8Sb7mCGu97Uk08zMdUJaDKRR8tGl2yR86PmTTDiLCIM5wOJaACIVGWlYN5CAuVK93fLxvSBds8isxNJrZ/laGsJ8rxmiCihf+D0Q+YGnRj1Q7PSuNtk3Wblnypi2rbMgLjBQZLWpBgz96VVADiSMccK5oY4vPe4ujyvAWLRKZcAt1ijYcKfxRkbMCGiIo+AcLhK2i4uQmd6+5xRjkcGd9n5sAhNuyS9WIZ6KbAdYsrry/THE0Z+mbIBFWNVMmhUZ8fujML1BlI2A/BIdf3rndtSZpN0m9S+CthnWAzTNOXlaxPis+wEJm/7sX8w4o4FS71exKA8Ntm4J46/8s+sxnY2pVEX7li6q2ppQbNuACnxjyqqSA7watf87TdnMwYd1IrvRiOGskAlebT8GvJ6em2CkZW8Chl0R7UkYlioLFUeV5I/ElwbV5SvGXAIL6xX1wbwc4en6GBS47/XeIRYUMuRUutK1N6tOrLOAg/aq3Fll+hvfsafcHOkW8T1h9H8GhdjmTBaJLudQbXDuS2ykHY3PGXqBsuL4JwYOpBWgeXb3OCHPT2jtgD3JLHAJmvAzgV2budW31iU4+LWLuiQbTflX+dLBl7Cib1Q+d9bkGFyD4jm6W63G/1c5UVvthq0hU8E1vDwUxAH/W2GHYF1osQ==");
	//plainText = Clipperz.Crypto.AES.decrypt(key, cypherText).asString();
        //alert(plainText);
        
	is(plainText, "some text to encrypt");

	randomBytes = Clipperz.Crypto.PRNG.defaultRandomGenerator().getRandomBytes(256/8);
	hashedValue = Clipperz.Crypto.SHA.sha256(new Clipperz.ByteArray("text to hash"));

//#############################################################################
 
} catch (err) {
    
    var s = "test suite failure!\n";
    var o = {};
    var k = null;
    for (k in err) {
        // ensure unique keys?!
        if (!o[k]) {
            s +=  k + ": " + err[k] + "\n";
            o[k] = err[k];
        }
    }
    ok ( false, s );
}

</script>
</pre>
</body>
</html>
